[x86/Linux] Fix GetCallerSp (dotnet/coreclr#9384)
* [x86/Linux] Fix GetCallerSp

* Do NOT pop stack argument for TransitionFrame

* Add ResumeEsp to CONTEXT


Commit migrated from https://github.com/dotnet/coreclr/commit/00fc8f9f212bc00eda9c94707b50e278d9b8d7c4

+++ src/coreclr/src/vm/eetwain.cpp
@@ -3099,7 +3099,8 @@ size_t EECodeManager::GetCallerSp( PREGDISPLAY  pRD )
     {
         EnsureCallerContextIsValid(pRD, NULL);
     }
-    return (size_t) (GetSP(pRD->pCallerContext));
+
+    return GetSP(pRD->pCallerContext);
 }
 
 #endif // WIN64EXCEPTIONS && !CROSSGEN_COMPILE
